Moved iterator_cache and secondary_key_helper out of apply_context and into backend_store directory.
